Add kconfig dotconfig import crate
This crate provides dotconfig import mechanisme from build script.
KCONFIG entries are use as cfg flag is present.
For entry with type other than boolean, there are also exported as
env variable.

This crate provides also macro for retriveing entry value if not
boolean. Supported type are int, hex and string.
hex and int can be converted to integer if required.

For boolean value use cfg! rust built-in macro (may be provide a wraper
for this in future release.